--- a/lib/sidekiq/metrics/tracking.rb
+++ b/lib/sidekiq/metrics/tracking.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "time"
 require "sidekiq"
 require "sidekiq/metrics/shared"