@mperham 2023-02-06
Looks really good. I'm not sure why `json_safe_recursive_hash` is so much faster than `json_safe_optimized`, they seem to be the exact same code, only using Procs instead of methods. Any idea?