@fatkodima 2023-02-06
For example,
```ruby
case x
when String, Integer
  do_something
when Float
  do_something_else
end
```

```
 $ ruby --dump=insns example.rb
== disasm: #<ISeq:<main>@example.rb:1 (1,0)-(8,3)> (catch: false)
0000 putself                                                          (   3)[Li]
0001 opt_send_without_block                 <calldata!mid:x, argc:0, FCALL|VCALL|ARGS_SIMPLE>
0003 opt_getconstant_path                   <ic:0 String>             (   4)
0005 topn                                   1
0007 opt_send_without_block                 <calldata!mid:===, argc:1, FCALL|ARGS_SIMPLE>
0009 branchif                               30
0011 opt_getconstant_path                   <ic:1 Integer>
0013 topn                                   1
0015 opt_send_without_block                 <calldata!mid:===, argc:1, FCALL|ARGS_SIMPLE>
0017 branchif                               30
0019 opt_getconstant_path                   <ic:2 Float>              (   6)
0021 topn                                   1
0023 opt_send_without_block                 <calldata!mid:===, argc:1, FCALL|ARGS_SIMPLE>
0025 branchif                               35
0027 pop                                                              (   3)
0028 putnil
0029 leave                                                            (   7)
0030 pop                                                              (   4)
0031 putself                                                          (   5)[Li]
0032 opt_send_without_block                 <calldata!mid:do_something, argc:0, FCALL|VCALL|ARGS_SIMPLE>
0034 leave                                                            (   7)
0035 pop                                                              (   6)
0036 putself                                                          (   7)[Li]
0037 opt_send_without_block                 <calldata!mid:do_something_else, argc:0, FCALL|VCALL|ARGS_SIMPLE>
0039 leave
```

We linearly scan all the items in each case, calling [`Module#===`](https://ruby-doc.org/core-3.1.1/Module.html#method-i-3D-3D-3D), which calls `kind_of?` under the hood, which checks if an object is an instance of a class or **its descendants**.

On the other hand, "recursive hash" uses `Hash#compare_by_identity`, which does not linear scan (as `case`) and compares just `object_id`s, which is fast.

The `case` in PR is pretty large (especially many items in the first branch), so here is the worse performance, I believe.