@mperham 2023-02-06
Cool, can you update with the hash solution? I like that.